+++ Src/FluentAssertions/Equivalency/Steps/EnumerableEquivalencyValidator.cs
@@ -16,12 +16,12 @@ internal class EnumerableEquivalencyValidator
 
     #region Private Definitions
 
-    private readonly IEquivalencyValidator parent;
+    private readonly IValidateChildNodeEquivalency parent;
     private readonly IEquivalencyValidationContext context;
 
     #endregion
 
-    public EnumerableEquivalencyValidator(IEquivalencyValidator parent, IEquivalencyValidationContext context)
+    public EnumerableEquivalencyValidator(IValidateChildNodeEquivalency parent, IEquivalencyValidationContext context)
     {
         this.parent = parent;
         this.context = context;
@@ -183,7 +183,7 @@ internal class EnumerableEquivalencyValidator
             unmatchedSubjectIndexes.RemoveAt(indexToBeRemoved);
         }
 
-        foreach (string failure in results.SelectClosestMatchFor(expectationIndex))
+        foreach (string failure in results.GetTheFailuresForTheSetWithTheFewestFailures(expectationIndex))
         {
             AssertionScope.Current.AddPreFormattedFailure(failure);
         }
@@ -195,7 +195,7 @@ internal class EnumerableEquivalencyValidator
     {
         using var scope = new AssertionScope();
 
-        parent.RecursivelyAssertEquality(new Comparands(subject, expectation, typeof(T)),
+        parent.AssertEquivalencyOf(new Comparands(subject, expectation, typeof(T)),
             context.AsCollectionItem<T>(expectationIndex));
 
         return scope.Discard();
@@ -207,7 +207,7 @@ internal class EnumerableEquivalencyValidator
         object subject = subjects[expectationIndex];
         IEquivalencyValidationContext equivalencyValidationContext = context.AsCollectionItem<T>(expectationIndex);
 
-        parent.RecursivelyAssertEquality(new Comparands(subject, expectation, typeof(T)), equivalencyValidationContext);
+        parent.AssertEquivalencyOf(new Comparands(subject, expectation, typeof(T)), equivalencyValidationContext);
 
         bool failed = scope.HasFailures();
         return !failed;