+++ Src/FluentAssertions/Equivalency/Steps/XElementEquivalencyStep.cs
@@ -5,13 +5,13 @@ namespace FluentAssertions.Equivalency.Steps;
 public class XElementEquivalencyStep : EquivalencyStep<XElement>
 {
     protected override EquivalencyResult OnHandle(Comparands comparands, IEquivalencyValidationContext context,
-        IEquivalencyValidator nestedValidator)
+        IValidateChildNodeEquivalency nested)
     {
         var subject = (XElement)comparands.Subject;
         var expectation = (XElement)comparands.Expectation;
 
         subject.Should().BeEquivalentTo(expectation, context.Reason.FormattedMessage, context.Reason.Arguments);
 
-        return EquivalencyResult.AssertionCompleted;
+        return EquivalencyResult.EquivalencyProven;
     }
 }